refactor(app): moves MapCard to shared components folder
